+++ checkpoint/orbax/checkpoint/json_checkpoint_handler_test.py
@@ -13,6 +13,11 @@
 # limitations under the License.
 
 """Tests for JsonCheckpointHandler."""
+
+import asyncio
+import time
+from typing import Optional
+
 from absl import flags
 from absl.testing import absltest
 from etils import epath
@@ -29,28 +34,66 @@ JsonSaveArgs = json_checkpoint_handler.JsonSaveArgs
 JsonRestoreArgs = json_checkpoint_handler.JsonRestoreArgs
 
 
+class TestJsonCheckpointHandler(JsonCheckpointHandler):
+
+  def __init__(
+      self,
+      filename: Optional[str] = None,
+  ):
+    super().__init__(filename=filename)
+    self._handler = json_checkpoint_handler.JsonCheckpointHandler
+    self._filename = filename or 'metadata'
+
+  def _save_fn(self, x, directory):
+    time.sleep(5)
+    return super()._save_fn(x, directory)
+
+  def close(self):
+    self._executor.shutdown(wait=True)
+
+
 class JsonCheckpointHandlerTest(absltest.TestCase):
 
   def setUp(self):
     super().setUp()
     self.directory = epath.Path(
-        self.create_tempdir(name='checkpointing_test').full_path)
+        self.create_tempdir(name='checkpointing_test').full_path
+    )
 
   def test_save_restore(self):
     item = {'a': 1, 'b': {'c': 'test1', 'b': 'test2'}, 'd': 5.5}
-    checkpointer = JsonCheckpointHandler()
-    checkpointer.save(self.directory, args=JsonSaveArgs(item))
-    restored = checkpointer.restore(self.directory)
+    handler = JsonCheckpointHandler()
+    handler.save(self.directory, args=JsonSaveArgs(item))
+    restored = handler.restore(self.directory)
     self.assertEqual(item, restored)
 
   def test_save_restore_filename(self):
     item = {'a': 1, 'b': {'c': 'test1', 'b': 'test2'}, 'd': 5.5}
-    checkpointer = JsonCheckpointHandler(filename='file')
-    checkpointer.save(self.directory, args=JsonSaveArgs(item))
-    restored = checkpointer.restore(self.directory)
+    handler = JsonCheckpointHandler(filename='file')
+    handler.save(self.directory, args=JsonSaveArgs(item))
+    restored = handler.restore(self.directory)
     self.assertEqual(item, restored)
     self.assertTrue((self.directory / 'file').exists())
     self.assertFalse((self.directory / 'metadata').exists())
+
+  def test_async_save_restore(self):
+    item = {'a': 1, 'b': {'c': 'test1', 'b': 'test2'}, 'd': 5.5}
+    handler = TestJsonCheckpointHandler()
+
+    async def run_async_test():
+      async_futures = await handler.async_save(
+          self.directory, args=JsonSaveArgs(item)
+      )
+      # item wasn't saved at directory right away.
+      self.assertFalse((self.directory / 'metadata').exists())
+      for future in async_futures:
+        future.result()
+      self.assertTrue((self.directory / 'metadata').exists())
+      restored = handler.restore(directory=self.directory, args=None)
+      self.assertEqual(item, restored)
+
+    asyncio.run(run_async_test())
+    handler.close()
 
 
 if __name__ == '__main__':